requirements.txt
asgiref==3.2.3
Django==2.2.7
django-appconf==1.0.4
django-cleanup==4.0.0
django-modeltranslation==0.15
easy-thumbnails==2.7
mysqlclient==1.4.6
Pillow==7.0.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
